Find more submit buttons
The only [type=submit] that is legal is a button element. And a button
element with a missing type attribute or a value for its type attribute
that is not “reset” or “button” is defined as being a submit button:

https://html.spec.whatwg.org/multipage/form-elements.html#attr-button-type

I have encountered at least three login forms that use as their submit
buttons a <button> element with no or an empty type attribute; these
were not formerly caught by this extension, making it just call
form.submit() instead, which was in each case the wrong thing to do.
With this change, they should work fine.